Torch float16

+++ app/run-llama.py
@@ -54,14 +54,12 @@ def classify_sentiment(prompt):
   return sentiment,total_time
 
 if device=='xla':
-  model = NeuronModelForCausalLM.from_pretrained(compiled_model_id)
+  model = NeuronModelForCausalLM.from_pretrained(compiled_model_id, torch_dtype=torch.float16)
 elif device=='cuda': 
   # model = AutoModelForCausalLM.from_pretrained(model_id,use_cache=True,device_map='auto',torch_dtype=torch.float16,quantization_config=quantization_config,)
-  model = AutoModelForCausalLM.from_pretrained(model_id)
+  model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float16)
 elif device=='cpu':
-  # model=AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float16)
-  model = AutoModelForCausalLM.from_pretrained(model_id)
-  # model = AutoModelForCausalLM.from_pretrained(model_id,use_cache=True,device_map='auto',torch_dtype=torch.float16,quantization_config=quantization_config,)
+  model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float16)
 
 gentext("write a poem")
